requirements.txt
fastapi==0.79.0
pydantic==1.9.0
uvicorn==0.18.2
python-dotenv==0.19.2
validators==0.18.2
sqlalchemy==2.0.17
asyncpg==0.26.0
psycopg2-binary==2.9.3

orjson==3.9.1
alembic==1.11.1
httpx==0.24.1
pytest==7.4.0
python-multipart==0.0.6
psycopg2==2.9.6

PyJWT==2.6.0
python-decouple==3.7
uvloop==0.17.0